requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.2.0
openai==0.28.1
propcache==0.3.1
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
rich==14.0.0
simple-websocket==1.1.0
sniffio==1.3.1
tqdm==4.67.1
twilio==9.5.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
